fix: create hooks for properties
